+++ services/users/infra/storage/user_repository.go
@@ -1,0 +1,13 @@
+package storage
+
+import "github.com/italoservio/braz_ecommerce/packages/database"
+
+type UserRepositoryInterface interface{}
+
+type UserRepositoryImpl struct {
+	database *database.Database
+}
+
+func NewUserRepositoryImpl(db *database.Database) *UserRepositoryImpl {
+	return &UserRepositoryImpl{database: db}
+}